PER-153 rename adapter and set new version Refactored the adapter to kinetic_platform after a conversation with James and Ben. The idea behind the rename is that we will end up with one adapter to interact with all Kinetic platform components. Versioning adapter (5.0.0-snapshot) to match new platform version.
