+++ bridge-adapters/kinetic-bridgehub-adapter-kinetic-platform/lib/kinetic-bridgehub-adapter-1.0.3.jar
