+++ bridge-adapters/kinetic-bridgehub-adapter-kinetic-platform/src/main/resources/META-INF/adapters/com.kineticdata.bridgehub.adapter.BridgeAdapter
@@ -1,0 +1,1 @@
+com.kineticdata.bridgehub.adapter.kinetic.platform.KineticCoreAdapter
