+++ bridge-adapters/kinetic-bridgehub-adapter-kinetic-platform/src/main/resources/com.kineticdata.bridgehub.adapter.kinetic.platform.KineticCoreAdapter.version
