+++ bridge-adapters/kinetic-bridgehub-adapter-kinetic-platform/src/test/resources/bridge-config.yml
