--- bridge-adapters/kinetic-bridgehub-adapter-kineticcore-v2/pom.xml
@@ -1,176 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-
-  <groupId>com.kineticdata.bridges.adapter</groupId>
-  <artifactId>kinetic-bridgehub-adapter-kineticcore-v2</artifactId>
-  <version>2.0.0</version>
-  <packaging>jar</packaging>
-
-  <name>kinetic-bridgehub-adapter-kineticcore-v2</name>
-  <scm>
-      <connection>scm:git:https://github.com/kineticcommunity/kinetic-bridgehub-adapter-kineticcore.git</connection>
-      <developerConnection>scm:git:https://github.com/kineticcommunity/kinetic-bridgehub-adapter-kineticcore.git</developerConnection>
-  </scm>
-  <url>http://maven.apache.org</url>
-
-  <distributionManagement>
-      <repository>
-          <id>com.kineticdata</id>
-          <url>http://utilities.kineticdata.com/nexus/content/repositories/releases/</url>
-      </repository>
-      <snapshotRepository>
-          <id>com.kineticdata.snapshots</id>
-          <url>http://utilities.kineticdata.com/nexus/content/repositories/snapshots</url>
-      </snapshotRepository>
-  </distributionManagement>
-
-  <repositories>
-      <repository>
-        <id>com.kineticdata</id>
-        <url>http://utilities.kineticdata.com/nexus/content/repositories/releases/</url>
-      </repository>
-  </repositories>
-
-  <properties>
-    <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-    <maven.compiler.source>1.8</maven.compiler.source>
-    <maven.compiler.target>1.8</maven.compiler.target>
-    <maven.build.timestamp.format>yyyy-MM-dd HH:mm</maven.build.timestamp.format>
-  </properties>
-
-  <dependencies>
-    <dependency>
-        <groupId>com.kineticdata.bridges.adapter</groupId>
-        <artifactId>kinetic-bridgehub-adapter</artifactId>
-        <version>1.2.0</version>
-    </dependency>
-    <dependency>
-      <groupId>junit</groupId>
-      <artifactId>junit</artifactId>
-      <version>4.11</version>
-      <scope>test</scope>
-      <type>jar</type>
-    </dependency>
-    <dependency>
-      <groupId>org.apache.httpcomponents</groupId>
-      <artifactId>httpclient</artifactId>
-      <version>4.5.1</version>
-    </dependency>
-    <dependency>
-      <groupId>org.yaml</groupId>
-      <artifactId>snakeyaml</artifactId>
-      <version>1.16</version>
-      <scope>test</scope>
-      <type>jar</type>
-    </dependency>
-    <dependency>
-        <groupId>org.slf4j</groupId>
-        <artifactId>slf4j-api</artifactId>
-        <version>1.7.10</version>
-    </dependency>
-    <dependency>
-        <groupId>com.googlecode.json-simple</groupId>
-        <artifactId>json-simple</artifactId>
-        <version>1.1.1</version>
-    </dependency>
-    <dependency>
-        <groupId>commons-lang</groupId>
-        <artifactId>commons-lang</artifactId>
-        <version>2.6</version>
-    </dependency>
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
-      <version>18.0</version>
-    </dependency>
-    <dependency>
-      <groupId>org.hamcrest</groupId>
-      <artifactId>hamcrest-all</artifactId>
-      <version>1.3</version>
-      <scope>test</scope>
-      <type>jar</type>
-    </dependency>
-    <dependency>
-      <groupId>org.hamcrest</groupId>
-      <artifactId>hamcrest-core</artifactId>
-      <version>1.3</version>
-      <type>jar</type>
-    </dependency>
-  </dependencies>
-
-  <build>
-      <resources>
-          <resource>
-              <directory>src/main/resources</directory>
-              <filtering>true</filtering>
-          </resource>
-      </resources>
-      <plugins>
-          <plugin>
-                <groupId>org.codehaus.mojo</groupId>
-                <artifactId>buildnumber-maven-plugin</artifactId>
-                <version>1.3</version>
-                <executions>
-                    <execution>
-                        <phase>validate</phase>
-                        <goals>
-                            <goal>create</goal>
-                        </goals>
-                        <configuration>
-                            <doCheck>false</doCheck>
-                            <doUpdate>false</doUpdate>
-                            <shortRevisionLength>7</shortRevisionLength>
-                        </configuration>
-                    </execution>
-                    <execution>
-                        <id>generate-timestamp</id>
-                        <phase>validate</phase>
-                        <goals>
-                            <goal>create</goal>
-                        </goals>
-                        <configuration>
-                            <doCheck>false</doCheck>
-                            <doUpdate>false</doUpdate>
-                            <buildNumberPropertyName>buildDate</buildNumberPropertyName>
-                            <format>{0,date,yyyy-MM-dd HH:mm:ss Z}</format>
-                            <items>
-                                <item>timestamp</item>
-                            </items>
-                        </configuration>
-                    </execution>
-                </executions>
-            </plugin>
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-jar-plugin</artifactId>
-                <version>3.0.2</version>
-                <configuration>
-                    <archive>
-                        <manifestEntries>
-                            <Build-Time>${maven.build.timestamp}</Build-Time>
-                        </manifestEntries>
-                    </archive>
-                </configuration>
-            </plugin>
-            <!-- Testing -->
-            <plugin>
-                <groupId>org.apache.maven.plugins</groupId>
-                <artifactId>maven-surefire-plugin</artifactId>
-                <version>2.11</version>
-                <dependencies>
-                    <dependency>
-                        <groupId>org.apache.maven.surefire</groupId>
-                        <artifactId>surefire-junit47</artifactId>
-                        <version>2.12</version>
-                    </dependency>
-                </dependencies>
-                <configuration>
-                    <!-- Allow the project to be built even if tests are failing. -->
-                    <testFailureIgnore>true</testFailureIgnore>
-                </configuration>
-            </plugin>
-        </plugins>
-        <finalName>${project.artifactId}-${project.version}</finalName>
-    </build>
-</project>
