--- bridge-adapters/kinetic-bridgehub-adapter-kineticcore-v2/src/main/resources/META-INF/adapters/com.kineticdata.bridgehub.adapter.BridgeAdapter
@@ -1,1 +1,0 @@
-com.kineticdata.bridgehub.adapter.kineticcore.v2.KineticCoreAdapter
